requirements.txt
dbt-core==1.7.7
dbt-postgres==1.7.7
acryl-datahub==0.12.1.4
astronomer-cosmos==1.3.2
apache-airflow-providers-postgres==5.10.0

